[regression] initial version regdb.py
The purpose of this tool is to more easily query and list the test cases and
some of their specifics in order to get an overview. For instance:

	$ ./regdb.py -v

will list all known (and enabled) test cases including what the tool calls
"flags" for each one. Shortened example output:

	esbmc-cpp11/constructors/Constructor14-1                               : cxx
	esbmc-cpp11/constructors/Constructor15-fail                            : bug cxx
	esbmc-unix/00_race06                                                   : c
	esbmc-unix/00_race07                                                   : 32 c
	esbmc-unix/github_1147                                                 : cxx
	esbmc-unix/github_28_successful                                        : 64 c
	Interval-analysis-ibex-contractor/01-arithmetic-operations-fail        : c ia
	jimple/kt-array-false                                                  : jimple
	mathsat/00_Division_false-no-overflow                                  : c mathsat

The list of flags can be obtained with -l.